fix(接口测试): 修复资源池执行报错问题
--bug=1016208 --user=赵勇 【接口测试】批量执行，选择local/node资源池，都无法执行 https://www.tapd.cn/55049933/s/1231721

+++ src/main/java/io/metersphere/api/jmeter/utils/FileUtils.java
@@ -54,7 +54,9 @@ public class FileUtils {
 
         File[] files = getFiles(file);
         for (int i = 0; i < files.length; i++) {
-            files[i].delete();
+            if (files[i].exists()) {
+                files[i].delete();
+            }
         }
     }
 }
